% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EQ5D-as.data.frame.R
\name{as.data.frame.EQ5D}
\alias{as.data.frame.EQ5D}
\title{Coerce an EQ5D object to a data frame}
\usage{
\method{as.data.frame}{EQ5D}(x, row.names, optional, ...)
}
\arguments{
\item{x}{An EQ5D object.}

\item{row.names}{Not currently used.}

\item{optional}{Not currently used.}

\item{...}{Not currently used.}
}
\value{
The data frame underlying the EQ5D object. Only column names are preserved
with all other attributes, including row names, dropped.
}
\description{
Method to convert an EQ5D object to a data frame.
}
\note{
Apart from \code{x} no other parameters are used and they are only present
in the method signature for compatibility with the underlying generic.
}
