% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EQ5D-class.R
\name{as_eq5d}
\alias{as_eq5d}
\alias{as_eq5d5l}
\alias{as_eq5d3l}
\alias{as_eq5dy3l}
\title{Coerce to an EQ5D object}
\usage{
as_eq5d5l(
  x,
  respondentID,
  surveyID,
  mobility,
  self_care,
  usual,
  pain,
  anxiety,
  vas,
  ...
)

as_eq5d3l(
  x,
  respondentID,
  surveyID,
  mobility,
  self_care,
  usual,
  pain,
  anxiety,
  vas,
  ...
)

as_eq5dy3l(
  x,
  respondentID,
  surveyID,
  mobility,
  self_care,
  usual,
  pain,
  anxiety,
  vas,
  ...
)
}
\arguments{
\item{x}{A data frame like object.}

\item{respondentID}{\verb{[character]} The name of a variable in \code{x} that uniquely
identifies respondents.}

\item{surveyID}{\verb{[character]} Name of variable \code{x} that uniquely identifies
surveys over time.}

\item{mobility}{\verb{[character]} Name of the 'mobility' dimension in \code{x}.}

\item{self_care}{\verb{[character]} Name of the 'self-care' dimension in \code{x}.}

\item{usual}{\verb{[character]} Name of the 'usual activities' dimension in \code{x}.}

\item{pain}{\verb{[character]} Name of the 'pain / discomfort' dimension in \code{x}.}

\item{anxiety}{\verb{[character]} Name of the 'anxiety / depression' dimension in
\code{x}.}

\item{vas}{\verb{[character]} Name of the 'visual analogue score' variable in \code{x}.}

\item{...}{Not currently used.}
}
\value{
A \link[tibble:tbl_df-class]{tibble} like \verb{<EQ5D3L>}, \verb{<EQ5D5L>} or \verb{<EQ5DY3L>}
object.
}
\description{
We define an \verb{<EQ5D>} object as a data frame meeting the following criteria:
\itemize{
\item It contains columns that represent dimensions from the
\href{https://euroqol.org/information-and-support/euroqol-instruments/}{EQ5D survey}
specification as well as a column representing the Visual Analogue Score.
\item It contains a column that acts as a unique respondent identifier and
another that identifies different surveys over time. Together these
should uniquely identify a response and no combination should be
duplicated within the data frame.
}

\verb{<EQ5D3L>}, \verb{<EQ5D5L>} and \verb{<EQ5DY3L>} objects are defined as a subclass of
EQ5D objects with the additional restriction  that the corresponding
dimension columns in \code{x} are either NA or whole numbers bounded below by 1
and above by 3 or 5 (depending on the survey type).
}
\examples{

data(eq5d3l_example)
as_eq5d3l(
    eq5d3l_example,
    respondentID = "respondentID",
    surveyID = "surveyID",
    mobility = "MO",
    self_care = "SC",
    usual = "UA",
    pain = "PD",
    anxiety = "AD",
    vas = "vas"
)

}
