% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-class.R
\name{as_utility}
\alias{as_utility}
\title{Coerce to a utility object}
\usage{
as_utility(x, respondentID, surveyID, country, type, value)
}
\arguments{
\item{x}{\verb{[data.frame]}.}

\item{respondentID}{\verb{[character]} Unique respondent identifier. The name of a
variable in \code{x} that uniquely identifies respondents.}

\item{surveyID}{\verb{[character]} Name of variable in \code{x} that uniquely
identifies surveys over time. To avoid ambiguity the specified variable must
be either numeric or a factor (in which case the order will be taken as that
given by the factor levels).}

\item{country}{\verb{[character]} Name of variable in \code{x} representing the utility
country.}

\item{type}{\verb{[character]} Name of variable in \code{x} representing the utility
country.}

\item{value}{\verb{[character]} Name of variable in \code{x} representing the utility
value.}
}
\value{
A \link[tibble:tbl_df-class]{tibble} like \verb{<utility>} object.
}
\description{
Coerce data frames to utility objects

A utility object contains is a data frame that meets the following criteria:
\itemize{
\item It contains a column that acts as a unique respondent identifier and
another that identifies different surveys over time.
\item It contains additional columns that represent the country, type and value
of a utility that has previously been calculated.
\item Together, Each combination of respondent identifier, survey identifier,
utility country and utility type should be unique and not duplicated
across rows.
}
}
