% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_limitation.R
\name{calculate_limitation}
\alias{calculate_limitation}
\alias{calculate_limitation.default}
\alias{calculate_limitation.EQ5D}
\title{Calculate limitation}
\usage{
calculate_limitation(x, ...)

\method{calculate_limitation}{default}(x, ...)

\method{calculate_limitation}{EQ5D}(x, ...)
}
\arguments{
\item{x}{An \R object.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A \link[tibble:tbl_df-class]{tibble} of the fraction of individuals without
limitation.
}
\description{
Generic function that calculates the limitation of survey responses across
response dimensions. Here we define a limitation to be a dimension value not
equal to 1. Methods are provided for \link[=as_eq5d]{EQ5D} objects.
}
\examples{

data(eq5d3l_example)
dat <- as_eq5d3l(
    eq5d3l_example,
    respondentID = "respondentID",
    surveyID = "surveyID",
    mobility = "MO",
    self_care = "SC",
    usual = "UA",
    pain = "PD",
    anxiety = "AD",
    vas = "vas"
)
calculate_limitation(dat)

}
