% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_pchc.R
\name{calculate_pchc}
\alias{calculate_pchc}
\title{Calculate the Paretian Classification of Health Change}
\usage{
calculate_pchc(pre, post, no.problems = TRUE, by.dimension = FALSE)
}
\arguments{
\item{pre}{An EQ5D object for a single survey.}

\item{post}{An EQ5D object for a single survey.}

\item{no.problems}{boolean. Summarise 11111 "No change" subjects in a "No problems" 
group.}

\item{by.dimension}{boolean. Summarise results by each EQ-5D dimension rather 
than by the whole dataset.}
}
\value{
For \code{by.dimension = FALSE}:
\itemize{
\item A data frame with columns 'Change', 'Number' and 'Percent'.
}

For \code{by.dimension = TRUE}:
\itemize{
\item A data frame with columns 'Dimension', 'Change', 'Number' and 'Percent'.
}
}
\description{
\code{calculate_pchc()} calculates the Paretian Classification of Health Change
(PCHC) in an individuals health state between two surveys. It wraps the
\code{eq5d::pchc()} function providing methods for \link[=as_eq5d]{EQ5D} objects.
}
\examples{

data(eq5d3l_example)
dat <- as_eq5d3l(
    eq5d3l_example,
    respondentID = "respondentID",
    surveyID = "surveyID",
    mobility = "MO",
    self_care = "SC",
    usual = "UA",
    pain = "PD",
    anxiety = "AD",
    vas = "vas"
)
grp1 <- subset(dat, Group == "Group1")
grp2 <- subset(dat, Group == "Group2")
calculate_pchc(grp1, grp2)
calculate_pchc(grp1, grp2, by.dimension = TRUE)

}
