% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_utility.R
\name{calculate_utility}
\alias{calculate_utility}
\alias{calculate_utility.default}
\alias{calculate_utility.EQ5D5L}
\alias{calculate_utility.EQ5D3L}
\alias{calculate_utility.EQ5DY3L}
\alias{add_utility}
\alias{add_utility.default}
\alias{add_utility.EQ5D5L}
\alias{add_utility.EQ5D3L}
\alias{add_utility.EQ5DY}
\title{Calculate utility}
\usage{
calculate_utility(x, type, country, ...)

\method{calculate_utility}{default}(x, type, country, ...)

\method{calculate_utility}{EQ5D5L}(
  x,
  type = "VT",
  country = "England",
  drop = TRUE,
  age = NULL,
  sex = NULL,
  ...
)

\method{calculate_utility}{EQ5D3L}(
  x,
  type = "VT",
  country = "England",
  drop = TRUE,
  age = NULL,
  sex = NULL,
  ...
)

\method{calculate_utility}{EQ5DY3L}(
  x,
  type = "VT",
  country = "England",
  drop = TRUE,
  age = NULL,
  sex = NULL,
  ...
)

add_utility(x, type, country, ...)

\method{add_utility}{default}(x, type, country, ...)

\method{add_utility}{EQ5D5L}(x, type = "VT", country = "England", age = NULL, sex = NULL, ...)

\method{add_utility}{EQ5D3L}(x, type = "VT", country = "England", age = NULL, sex = NULL, ...)

\method{add_utility}{EQ5DY}(x, type = "VT", country = "England", age = NULL, sex = NULL, ...)
}
\arguments{
\item{x}{An \R object.}

\item{type}{\verb{[character]}

Method type(s) used for calculating the value sets.

For EQ5D3L inputs this can be:
\itemize{
\item "TTO", the time trade-off valuation technique;
\item "VAS", the visual analogue scale valuation technique;
\item "RCW", a reverse crosswalk conversion to EQ5D5L values; or
\item "DSU", the NICE Decision Support Unit's model that allows mappings on to
EQ5D5L values accounting for both age and sex.
}

For EQ5D5L inputs this can be:
\itemize{
\item "VT",  value sets generated via a EuroQol standardised valuation study
protocol;
\item "CW",  a crosswalk conversion EQ5D3L values; or
\item "DSU", the NICE Decision Support Unit's model that allows mappings on to
EQ5D5L values accounting for both age and sex.
}}

\item{country}{\verb{[character]}

Value set countries to use.}

\item{...}{Further arguments passed to or from other methods.}

\item{drop}{\verb{[logical]}

If TRUE (default), only columns corresponding to the surveyID and
respondentID are kept from the input \code{x}.}

\item{age}{\verb{[character]}

Column in \code{x} representing the age, in years, of the respondent.

Only used if \code{type = "DSU"}.}

\item{sex}{\verb{[character]}

Column in \code{x} representing the sex, in years, of the respondent.

Only used if \code{type = "DSU"}.

Column entries must be one of "Male", "M", "Female" or "F" (case insensitive).}
}
\value{
A data frame of utility values linked to responses.
}
\description{
Generic function that calculates EQ5D index scores for given value sets.
}
\details{
\itemize{
\item \code{calculate_utility()} returns the utility index scores for a given object
and value set combination. Methods that wrap functionality of the
\href{https://cran.r-project.org/package=eq5d}{eq5d} package are provided for
EQ5D objects.
\item \code{add_utility()} is a wrapper around \code{calculate_utility()} which keeps all
columns of the input data \code{x}.
}
}
\note{
The methods for \link[=as_eq5d]{EQ5D} objects expect \code{type} and \code{country} to be of
the same length but will recycle those of length one to a common size.
}
\examples{

data(eq5d3l_example)
dat <- as_eq5d3l(
    eq5d3l_example,
    respondentID = "respondentID",
    surveyID = "surveyID",
    mobility = "MO",
    self_care = "SC",
    usual = "UA",
    pain = "PD",
    anxiety = "AD",
    vas = "vas"
)
calculate_utility(dat, type = "TTO", country = c("UK", "Germany"))

}
\seealso{
\code{available_valuesets()} for the method / country combinations available for
each survey type.
}
