% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coercion.R
\name{coercion}
\alias{coercion}
\alias{as.data.frame.EQ5D}
\alias{as.data.table.EQ5D}
\alias{as_tibble.EQ5D}
\title{Coerce an EQ5D object}
\usage{
\method{as.data.frame}{EQ5D}(x, row.names, optional, ...)

\method{as.data.table}{EQ5D}(x, keep.rownames, ...)

\method{as_tibble}{EQ5D}(x, ..., .rows, .name_repair, rownames)
}
\arguments{
\item{x}{An EQ5D object.}

\item{row.names}{Not used.}

\item{optional}{Not used.}

\item{...}{Not used.}

\item{keep.rownames}{Not used.}

\item{.rows}{Not used.}

\item{.name_repair}{Not used.}

\item{rownames}{Not used.}
}
\value{
An appropriate representation of the data frame underlying the EQ5D object.
Only column names are preserved with all other attributes, including row
names, dropped.
}
\description{
Methods to convert an EQ5D object to a data frame, tibble or data.table.
}
\note{
Apart from \code{x} no other parameters are used and they are only present
in the method signatures for compatibility with the underlying generic.
}
