% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EQ5D5L_surveys}
\alias{EQ5D5L_surveys}
\title{EQ-5D-5L multiple survey example data}
\format{
A \link[tibble:tbl_df-class]{tibble} with 10000 rows and 11 variables:
\describe{
\item{surveyID}{Survey ID}
\item{respondentID}{Respondent ID}
\item{sex}{Respondent sex}
\item{age}{Respondent age}
\item{mobility}{Mobility dimension}
\item{self_care}{Self-care dimension}
\item{usual}{Usual activities dimension}
\item{pain}{Pain/discomfort dimension}
\item{anxiety}{Anxiety/depression dimension}
\item{vas}{VAS score}
\item{time_index}{Relative time within survey framework}
\item{dummy}{Random roolean variable}
}
}
\usage{
EQ5D5L_surveys
}
\description{
A dataset containing dimension values and vas score for 1000 respondents
across 10 surveys. Data was synthetically generated.
}
\keyword{datasets}
