% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EQ5D-new_eq5d.R
\name{new_eq5d}
\alias{new_eq5d}
\alias{new_eq5d3l}
\alias{new_eq5d5l}
\alias{new_eq5dy}
\alias{validate_eq5d}
\alias{validate_eq5d3l}
\alias{validate_eq5d5l}
\alias{validate_eq5dy}
\title{EQ5D constructors and validators}
\usage{
new_eq5d3l(
    x,
    respondentID, surveyID, time_index,
    mobility, self_care, usual, pain, anxiety,
    vas
)

new_eq5d3l(
    x,
    respondentID, surveyID, time_index,
    mobility, self_care, usual, pain, anxiety,
    vas
)

new_eq5dy(
    x,
    respondentID, surveyID, time_index,
    mobility, self_care, usual, pain, anxiety,
    vas
)

validate_eq5d(x, version)

validate_eq5d3l(x)

validate_eq5d5l(x)

validate_eq5dy(x)
}
\arguments{
\item{x}{\verb{[data.frame]} EQ5D survey data.}

\item{respondentID}{\verb{[character]} The name of a variable in \code{x} that uniquely
identifies respondents.}

\item{surveyID}{\verb{[character]} Name of variable in \code{x} that uniquely
identifies surveys over time.

To avoid ambiguity the specified variable must be either numeric or a
factor (in which case the  order will be taken as that given  by the factor
levels).}

\item{time_index}{\verb{[character]} Name of variable in \code{x} representing the
relative time within the survey framework.}

\item{mobility}{\verb{[character]} Name of the 'mobility' dimension in \code{x}.}

\item{self_care}{\verb{[character]} Name of the 'self-care' dimension in \code{x}.}

\item{usual}{\verb{[character]} Name of the 'usual activities' dimension in \code{x}.}

\item{pain}{\verb{[character]} Name of the 'pain / discomfort' dimension in \code{x}.}

\item{anxiety}{\verb{[character]} Name of the 'anxiety / depression' dimension in
\code{x}.}

\item{vas}{\verb{[character]} Name of the 'visual analogue score' variable in \code{x}.}

\item{version}{\verb{[character]} The EQ5D version. One of "3L", "5L" or "Y".}
}
\value{
An EQ5D3L, EQ5D5L or EQ5D3Y object (invisibly for the corresponding
validation functions).
}
\description{
\code{new_eq5d3l()}, \code{new_eq5d5l()} and \code{new_eq5dy()} are developer facing
functions for the construction of EQ5D objects that perform minimal checking
of input data. They should normally be followed by a call to
\code{validate_eq5d()} to ensure assumptions about the underlying data are valid.
\subsection{The EQ5D class}{

We define an EQ5D object as a data frame that meets the following criteria:
\itemize{
\item It contains columns that represent dimensions from the EQ5D survey
specification as well as a column representing the Visual Analogue Score.
\item It contains a column that acts as a unique respondent identifier and
another that identifies different surveys over time. Together these
should uniquely identify a response and no combination should be
duplicated within the data frame.
\item Finally, it contains a column that provides the relative time of a
response within the survey framework.
}

EQ5D3L, EQ5D5L and EQ5DY objects are defined as a subclass of EQ5D objects
with the additional restriction  that the corresponding dimension columns in
\code{x} are either NA or whole numbers bounded below by 1 and above by 3 or 5
(depending on the survey type).
}
}
\examples{

data("EQ5D5L_surveys")
dat <- EQ5D5L_surveys
dat$survey <- factor(dat$survey)

res <- new_eq5d5l(dat,
    respondentID = "respondentID",
    surveyID = "surveyID",
    mobility = "mobility",
    self_care = "self_care",
    usual = "usual",
    pain = "pain",
    anxiety = "anxiety",
    vas = "vas",
    time_index = "time_index"
)

validate_eq5d(res, version = "5L")
validate_eq5d5l(res)
try(validate_eq5d3l(res))
try(validate_eq5d(res, version = "3L"))

}
\seealso{
\code{\link{as_eq5d}} for user-facing alternatives.
}
