% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-new_utility.R
\name{new_utility}
\alias{new_utility}
\alias{validate_utility}
\title{Utility constructor and validator}
\usage{
new_utility(x, respondentID, surveyID, time_index, country, type, value)

validate_utility(xx)
}
\arguments{
\item{x}{\verb{[data.frame]}.}

\item{respondentID}{\verb{[character]} Unique respondent identifier. The name of a
variable in \code{x} that uniquely identifies respondents.}

\item{surveyID}{\verb{[character]} Name of variable in \code{x} that uniquely
identifies surveys over time. To avoid ambiguity the specified variable must
be either numeric or a factor (in which case the order will be taken as that
given by the factor levels).}

\item{time_index}{\verb{[character]} Name of variable in \code{x} representing the
relative time within the survey framework.}

\item{country}{\verb{[character]} Name of variable in \code{x} representing the utility
country.}

\item{type}{\verb{[character]} Name of variable in \code{x} representing the utility
country.}

\item{value}{\verb{[character]} Name of variable in \code{x} representing the utility
value.}

\item{xx}{An \R Object to validate.}
}
\value{
A utility object (invisibly for \code{validate_utility()}).
}
\description{
Construct and validate utility objects.
\subsection{The utility class}{

A utility object contains is an object that meets the following criteria:
\itemize{
\item It contains a column that acts as a unique respondent identifier and
another that identifies different surveys over time.
\item It contains a column that provides the relative time of a response within
the survey framework.
\item It contains additional columns that represent the country, type and value
of a utility that has previously been calculated.
\item Together, Each combination of respondent identifier, survey identifier,
utility country and utility type should be unique and not duplicated
across rows.
}
}
}
\details{
\code{new_utility()} creates a utility object with minimal checking of the input
arguments. It is mainly for developer use and should normally be accompanied
by a call to \code{validate_utility()} which performs further checking of the
underlying data.
}
