% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EQ5D-class.R
\name{summary.EQ5D}
\alias{summary.EQ5D}
\title{Summary method for eq5d objects}
\usage{
\method{summary}{EQ5D}(object, ..., tidy = FALSE)
}
\arguments{
\item{object}{An EQ5D object.}

\item{...}{Not currently used.}

\item{tidy}{\verb{[logical]}

Should results be returned as single, "tidy", table in long format?

If \code{FALSE}, output will be a list of frequency / proportion tables split by
survey ID.}
}
\description{
Provides a summary of dimension values by survey ID.
}
