% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EQ5D-as_eq5d.R
\name{as_eq5d}
\alias{as_eq5d}
\alias{as_eq5d5l}
\alias{as_eq5d5l.default}
\alias{as_eq5d5l.tbl_df}
\alias{as_eq5d5l.data.table}
\alias{as_eq5d5l.data.frame}
\alias{as_eq5d3l}
\alias{as_eq5d3l.default}
\alias{as_eq5d3l.tbl_df}
\alias{as_eq5d3l.data.table}
\alias{as_eq5d3l.data.frame}
\alias{as_eq5dy}
\alias{as_eq5dy.default}
\alias{as_eq5dy.tbl_df}
\alias{as_eq5dy.data.table}
\alias{as_eq5dy.data.frame}
\title{Coerce to an EQ5D object}
\usage{
as_eq5d5l(x, ...)

\method{as_eq5d5l}{default}(x, ...)

\method{as_eq5d5l}{tbl_df}(
  x,
  respondentID = "respondentID",
  surveyID = "surveyID",
  time_index = "time_index",
  mobility = "mobility",
  self_care = "self_care",
  usual = "usual",
  pain = "pain",
  anxiety = "anxiety",
  vas = "vas",
  drop = FALSE,
  ...
)

\method{as_eq5d5l}{data.table}(
  x,
  respondentID = "respondentID",
  surveyID = "surveyID",
  time_index = "time_index",
  mobility = "mobility",
  self_care = "self_care",
  usual = "usual",
  pain = "pain",
  anxiety = "anxiety",
  vas = "vas",
  drop = FALSE,
  ...
)

\method{as_eq5d5l}{data.frame}(
  x,
  respondentID = "respondentID",
  surveyID = "surveyID",
  time_index = "time_index",
  mobility = "mobility",
  self_care = "self_care",
  usual = "usual",
  pain = "pain",
  anxiety = "anxiety",
  vas = "vas",
  drop = FALSE,
  ...
)

as_eq5d3l(x, ...)

\method{as_eq5d3l}{default}(x, ...)

\method{as_eq5d3l}{tbl_df}(
  x,
  respondentID = "respondentID",
  surveyID = "surveyID",
  time_index = "time_index",
  mobility = "mobility",
  self_care = "self_care",
  usual = "usual",
  pain = "pain",
  anxiety = "anxiety",
  vas = "vas",
  drop = FALSE,
  ...
)

\method{as_eq5d3l}{data.table}(x, ...)

\method{as_eq5d3l}{data.frame}(
  x,
  respondentID = "respondentID",
  surveyID = "surveyID",
  time_index = "time_index",
  mobility = "mobility",
  self_care = "self_care",
  usual = "usual",
  pain = "pain",
  anxiety = "anxiety",
  vas = "vas",
  drop = FALSE,
  ...
)

as_eq5dy(x, ...)

\method{as_eq5dy}{default}(x, ...)

\method{as_eq5dy}{tbl_df}(
  x,
  respondentID = "respondentID",
  surveyID = "surveyID",
  time_index = "time_index",
  mobility = "mobility",
  self_care = "self_care",
  usual = "usual",
  pain = "pain",
  anxiety = "anxiety",
  vas = "vas",
  drop = FALSE,
  ...
)

\method{as_eq5dy}{data.table}(
  x,
  respondentID = "respondentID",
  surveyID = "surveyID",
  time_index = "time_index",
  mobility = "mobility",
  self_care = "self_care",
  usual = "usual",
  pain = "pain",
  anxiety = "anxiety",
  vas = "vas",
  drop = FALSE,
  ...
)

\method{as_eq5dy}{data.frame}(
  x,
  respondentID = "respondentID",
  surveyID = "surveyID",
  time_index = "time_index",
  mobility = "mobility",
  self_care = "self_care",
  usual = "usual",
  pain = "pain",
  anxiety = "anxiety",
  vas = vas,
  drop = FALSE,
  ...
)
}
\arguments{
\item{x}{An \R object.}

\item{...}{Further arguments passed to or from other methods.}

\item{respondentID}{\verb{[character]}

The name of a variable in \code{x} that uniquely identifies respondents.}

\item{surveyID}{\verb{[character]}

Name of variable \code{x} that uniquely identifies surveys over time.

To avoid ambiguity the specified variable should be either numeric or a
factor (in which case the  order will be taken as that given by the factor
levels).

A character variable in \code{x} will be accepted but converted, with
warning, via \code{as.factor()}.}

\item{time_index}{Name of variable in \code{x} representing the relative time within the survey
framework.

If the variable does not exist within \code{x} it will be created and set to
\code{NA_integer}.}

\item{mobility}{\verb{[character]}

Name of the 'mobility' dimension in \code{x}.}

\item{self_care}{\verb{[character]}

Name of the 'self-care' dimension in \code{x}.}

\item{usual}{\verb{[character]}

Name of the 'usual activities' dimension in \code{x}.}

\item{pain}{\verb{[character]}

Name of the 'pain / discomfort' dimension in \code{x}.}

\item{anxiety}{\verb{[character]}

Name of the 'anxiety / depression' dimension in \code{x}.}

\item{vas}{\verb{[character]}

Name of the 'visual analogue score' variable in \code{x}.}

\item{drop}{\verb{[logical]}

Should additional columns, not specified by arguments, be dropped.}
}
\value{
An EQ5D5L, EQ5D3L or EQ5DY object.
}
\description{
Generic for conversion to an \href{new_eq5d()}{EQ5D} object. Methods are provided
for data frame type objects.
\subsection{The EQ5D class}{

We define an EQ5D object as a data frame that meets the following criteria:
\itemize{
\item It contains columns that represent dimensions from the EQ5D survey
specification as well as a column representing the Visual Analogue Score.
\item It contains a column that acts as a unique respondent identifier and
another that identifies different surveys over time. Together these
should uniquely identify a response and no combination of these should
be duplicated within the data frame.
\item Finally, it contains a column that provides the relative time of a
response within the survey framework.
}

EQ5D3L, EQ5D5L and EQ5DY objects are defined as a subclass of EQ5D objects
with the additional restriction  that the corresponding dimension columns in
\code{x} are either NA or whole numbers bounded below by 1 and above by 3 or 5
(depending on the survey type).
}
}
\examples{

data("eq5d3l_example")
as_eq5d3l(eq5d3l_example)

}
\seealso{
\code{\link{new_eq5d}} for more minimal, developer focussed, alternatives.
}
