% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_valuesets.R
\name{available_valuesets}
\alias{available_valuesets}
\alias{available_valuesets.default}
\alias{available_valuesets.EQ5D5L}
\alias{available_valuesets.EQ5D3L}
\alias{available_valuesets.EQ5DY}
\alias{available_valuesets.character}
\title{Available value sets}
\usage{
available_valuesets(x, ...)

\method{available_valuesets}{default}(x, ...)

\method{available_valuesets}{EQ5D5L}(x, ...)

\method{available_valuesets}{EQ5D3L}(x, ...)

\method{available_valuesets}{EQ5DY}(x, ...)

\method{available_valuesets}{character}(x, ...)
}
\arguments{
\item{x}{An \R object.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A \link[tibble:tbl_df-class]{tibble} containing the available value sets for the
given object or survey type.
}
\description{
Generic function that lists value sets available.
}
\details{
\code{available_valuesets()} returns the available valuesets from the
\href{https://cran.r-project.org/package=eq5d}{eq5d} package. It is an s3
generic that wraps the \code{eq5d::valuesets()} function providing additional
methods for \code{\link[=as_eq5d]{eq5d}} and \code{character} objects.

For \code{character} objects, the input is expected to be the survey type with
various forms permitted:
\itemize{
\item "eq5d5l", "EQ5D5L", "eq-5d-5l", "EQ-5D-5L", "5L", "5l"
\item "eq5d3l", "EQ5D3L", "eq-5d-3l", "EQ-5D-3L", "3L", "3l"
\item "eq5dy3l" , "EQ5DY3L" , "eq-5d-y-3l" , "EQ-5D-Y-3L" , "Y3L" , "y3l"
}

If called with no arguments value sets for all available  version, type and
country combination are returned.
}
\examples{

data(eq5d3l_example)
dat <- as_eq5d3l(
    eq5d3l_example,
    respondentID = "respondentID",
    surveyID = "surveyID",
    mobility = "MO",
    self_care = "SC",
    usual = "UA",
    pain = "PD",
    anxiety = "AD",
    vas = "vas"
)
available_valuesets(dat)
available_valuesets("eq5d5l")

}
