% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_qalys.R
\name{calculate_qalys}
\alias{calculate_qalys}
\alias{calculate_qalys.default}
\alias{calculate_qalys.EQ5D}
\alias{calculate_qalys.utility}
\title{Calculate quality of life years}
\usage{
calculate_qalys(x, ...)

\method{calculate_qalys}{default}(x, ...)

\method{calculate_qalys}{EQ5D}(
  x,
  type,
  country,
  units = c("days", "weeks", "months", "quarters", "years"),
  baseline = NULL,
  ...
)

\method{calculate_qalys}{utility}(
  x,
  units = c("days", "weeks", "months", "quarters", "years"),
  baseline = NULL,
  ...
)
}
\arguments{
\item{x}{An \R object.}

\item{...}{Further arguments passed to or from other methods.}

\item{type}{\verb{[character]} Method type(s) used for calculating the value sets.

For EQ5D3L inputs this can be:
\itemize{
\item "TTO", the time trade-off valuation technique;
\item "VAS", the visual analogue scale valuation technique;
\item "RCW", a reverse crosswalk conversion to EQ5D5L values; or
\item "DSU", the NICE Decision Support Unit's model that allows mappings on to
EQ5D5L values accounting for both age and sex.
}

For EQ5D5L inputs this can be:
\itemize{
\item "VT",  value sets generated via a EuroQol standardised valuation study
protocol;
\item "CW",  a crosswalk conversion EQ5D3L values; or
\item "DSU", the NICE Decision Support Unit's model that allows mappings on to
EQ5D5L values accounting for both age and sex.
}}

\item{country}{\verb{[character]} Value set countries to use.}

\item{units}{\code{\link{character}} The units of the \code{time_index} column of \code{x}. Can
be one of "days", "weeks", "months", "quarters" or "years".}

\item{baseline}{(optional) Either a \code{character} string specifying the
surveyID, to use as a baseline or a data frame. If a data frame, it must have
at least two columns; one for the respondentID (with name matching that in
\code{x} input) and another (of any name) for the associated utility values. If
desired you can also specify columns that match on the utility and country
type columns of the input \code{x}.}
}
\value{
A data frame.
}
\description{
Generic for calculating quality of life years (QALY) metrics for
EQ5D survey respondents.
}
\details{
The methods provided for \code{\link[=new_utility]{utility}} and  \code{\link[=new_eq5d]{EQ5D}}
objects, return two metrics by default:
\itemize{
\item Firstly, a \code{raw} QALY. This is the area under the utility curve scaled to
the proportion of the year it corresponds to.
\item Secondly, a \code{loss_v_fullhealth} value. This represents the loss from
perfect health which is calculated by assuming all dimensions are 1
to calculate a full health QALY value.
}

Optionally, a third metric can also be returned, \code{loss_v_baseline}. This
represents the loss from a specified baseline utility value.
}
\note{
If a character string \code{baseline} argument is given then this must match
a surveyID to match against.  In this situation the survey \strong{is} still
included in the unadjusted, \code{raw}, calculation, prior to the calculation of
loss.

Alternatively the \code{baseline} argument can be specified as a data frame with a
column corresponding to the respondentID and another representing the
associated utility. Optionally columns corresponding to the utility country
and utility type can be included to allow more granular comparisons. For this
specification of baseline, it \strong{is not} included in the unadjusted, \code{raw},
calculation.
}
\examples{

data("EQ5D5L_surveys")
dat <- EQ5D5L_surveys
dat <- as_eq5d5l(dat)
calculate_qalys(dat, type = "VT", country = c("Denmark", "France"))

}
