% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eq5d3l_example}
\alias{eq5d3l_example}
\title{EQ-5D-3L example data}
\format{
A data frame with 200 rows and 6 variables:
\describe{
\item{respondentID}{Unique respondent identifier}
\item{surveyID}{Unique survey identifier}
\item{mobility}{Mobility dimension valuea}
\item{self_care}{Self-care dimension values}
\item{usual}{Usual activities dimension values}
\item{pain}{Pain/discomfort dimension values}
\item{anxiety}{Anxiety/depression dimension values}
\item{Group}{Observation group}
\item{vas}{VAS score}
\item{time_index}{Relative time within survey framework}
}
}
\source{
\url{https://github.com/fragla/eq5d/raw/master/inst/extdata/eq5d3l_example.xlsx}
}
\usage{
eq5d3l_example
}
\description{
A dataset containing dimensions and grouping for 200 observations. Data
was obtained from the \href{https://cran.r-project.org/package=eq5d}{eq5d}
package on 2021-09-16.
}
\keyword{datasets}
