% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eq5d3l_example}
\alias{eq5d3l_example}
\title{EQ-5D-3L example data}
\format{
A data frame with 200 rows and 6 variables:
\describe{
\item{respondentID}{Unique respondent identifier}
\item{surveyID}{Unique survey identifier}
\item{MO}{Mobility dimension valuea}
\item{SC}{Self-care dimension values}
\item{UA}{Usual activities dimension values}
\item{PD}{Pain/discomfort dimension values}
\item{AD}{Anxiety/depression dimension values}
\item{Group}{Observation group}
\item{vas}{VAS score}
\item{time}{Relative time within survey framework}
}
}
\source{
\url{https://github.com/fragla/eq5d/raw/master/inst/extdata/eq5d3l_example.xlsx}
}
\usage{
eq5d3l_example
}
\description{
A dataset containing dimensions and grouping for 200 observations. Data
was obtained from the \href{https://cran.r-project.org/package=eq5d}{eq5d}
package on 2022-06-15, with additional variables added for package
compatibility.
}
\keyword{datasets}
